pyproject.toml
[tool.poetry]
name = "llm-memorization1"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11.7"
transformers = "^4.46.2"
flwr = {version = "1.9.0", extras = ["simulation"]}
flwr-datasets = "^0.4.0"
hydra-core = "^1.3.2"
trl = "^0.8.6"
bitsandbytes = "^0.43.0"
scipy = "^1.13.0"
peft = "^0.6.2"
gradio = "^3.50.2"
scikit-learn = "^1.4.0"
cryptography = "^42.0.7"
setuptools = "^70.0.0"
wandb = "^0.17.1"
opacus = "^1.4.1"
torch = {url = "https://download.pytorch.org/whl/cu124/torch-2.5.1%2Bcu124-cp312-cp312-linux_x86_64.whl"}
datasets = "2.20.0"
omegaconf = "^2.3.0"
huggingface-hub = "^0.23.4"
ray = "^2.38.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



